refactor: eliminate UserAgentUtil code duplication
Create dedicated user_agent_utils.py module to consolidate UserAgentUtil
functionality that was duplicated between models.py and admin.py.

Changes:
- Create new src/django_audit_log/user_agent_utils.py module
- Remove duplicate UserAgentUtil class from models.py
- Remove duplicate UserAgentUtil class from admin.py
- Update imports in models.py, admin.py, utils.py, and tests.py
- Combine bot detection patterns from both versions
- Preserve Samsung device mapping functionality
- Maintain all existing functionality and API compatibility

Benefits:
- Single source of truth for user agent parsing logic
- Easier maintenance and updates
- Reduced code duplication (375 deletions, 239 insertions)
- All tests continue to pass
